gadia/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pillow==10.1.0
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
aiofiles==23.2.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
httpx==0.25.2
webdriver-manager==4.0.1
lxml==4.9.3 